utilities/hcm-db-migrations/requirements.txt
alembic==1.13.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
python-dotenv==1.0.0
